requirements.txt
appnope==0.1.0
decorator==4.0.10
falcon==1.0.0
gunicorn==19.6.0
httpie==0.9.4
ipython==5.0.0
ipython-genutils==0.1.0
oauthlib==1.1.2
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
python-mimeparse==1.5.2
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
